caching: column numeric histogram
