+++ runtime/queries/column_api_test.go
@@ -50,3 +50,17 @@ func BenchmarkColumnTimeGrain(b *testing.B) {
 		require.NotEmpty(b, q.Result)
 	}
 }
+
+func BenchmarkColumnNumericHistogram(b *testing.B) {
+	rt, instanceID := testruntime.NewInstanceForProject(b, "ad_bids")
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		q := &ColumnNumericHistogram{
+			TableName:  "ad_bids",
+			ColumnName: "bid_price",
+		}
+		err := q.Resolve(context.Background(), rt, instanceID, 0)
+		require.NoError(b, err)
+		require.NotEmpty(b, q.Result)
+	}
+}
